fix(laverna): fix cycle in type declarations
This basically just moves the types from `index.js` to `types.js` and renames `laverna.js` to `index.js`.

This could be more elegant if the sources were TS or even ESM.  😄
